+++ js/main.js
@@ -1,2 +1,2 @@
-// import './functions.js';
+// import './util.js';
 // import './mocks.js';
